Split utils code to other files
