+++ pkg/utils/conf_project.go
@@ -1,0 +1,6 @@
+package utils
+
+type ConfProject struct {
+	ProjectID string       `yaml:"projectID"`
+	Metrics   []ConfMetric `yaml:"metrics"`
+}
